+++ internal/oonirun/inputprocessor_test.go
@@ -1,4 +1,4 @@
-package engine
+package oonirun
 
 import (
 	"context"
